Convert Gemini model
