+++ convex/constants.ts
@@ -1,11 +1,11 @@
 export const AI_MODELS = [
   {
-    model: "gpt-4o",
-    name: "OpenAI - 4o Mini",
-  },
-  {
     model: "gemini-1.5-pro",
     name: "Gemini - 1.5 Pro",
+  },
+  {
+    model: "gpt-4o",
+    name: "OpenAI - 4o Mini",
   },
 ];
