+++ models/index.ts
@@ -1,3 +1,4 @@
+import { gemini15pro } from "./gemini-1.5-pro";
 import { gpt4o } from "./gpt-4o";
 
 export interface ModelResult {
@@ -10,6 +11,9 @@ export async function runModel(
   map: string[][],
 ): Promise<ModelResult> {
   switch (modelId) {
+    case "gemini-1.5-pro": {
+      return gemini15pro(map);
+    }
     case "gpt-4o": {
       return gpt4o(map);
     }
